--- packages/twenty-server/src/engine/metadata-modules/field-metadata/utils/compute-morph-relation-field-join-column-name.util.ts
@@ -1,13 +1,0 @@
-import { capitalize } from 'twenty-shared/utils';
-
-type ComputeMorphRelationFieldJoinColumnNameArgs = {
-  name: string;
-  targetObjectMetadataNameSingular: string;
-};
-
-export const computeMorphRelationFieldJoinColumnName = ({
-  name,
-  targetObjectMetadataNameSingular,
-}: ComputeMorphRelationFieldJoinColumnNameArgs) => {
-  return `${name}${capitalize(targetObjectMetadataNameSingular)}Id`;
-};
