+++ packages/twenty-shared/src/utils/fieldMetadata/compute-morph-relation-field-join-column-name.util.ts
@@ -11,3 +11,5 @@ export const computeMorphRelationFieldJoinColumnName = ({
 }: ComputeMorphRelationFieldJoinColumnNameArgs) => {
   return `${name}${capitalize(targetObjectMetadataNameSingular)}Id`;
 };
+
+// here
